# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aladiyat API",
    "version": "1.0.0",
    "description": "Backend API for horse app - auth, users, horses, favorites, bookings, services, videos.\n\nملاحظة: مسارات Express الفعلية في index.js هي مصدر الحقيقة؛ يُحدَّث هذا الملف يدوياً عند إضافة معاملات جديدة."
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Development" }
  ],
...
  "paths": {
...
    "/transport-requests/{requestId}/negotiation": {
      "post": {
        "tags": ["negotiation"],
        "summary": "Open negotiation for transport request (T4)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "requestId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Opened" }, "200": { "description": "Existing" } }
      },
      "get": {
        "tags": ["negotiation"],
        "summary": "Get negotiation with offers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "requestId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Negotiation view" }, "404": { "description": "Not found" } }
      }
    },
    "/transport-requests/{requestId}/offers": {
      "get": {
        "tags": ["negotiation"],
        "summary": "List offers for request",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "requestId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Offers list" } }
      },
      "post": {
        "tags": ["negotiation"],
        "summary": "Create offer or counter (append-only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "requestId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number" },
                  "currency": { "type": "string" },
                  "serviceId": { "type": "string" },
                  "parentOfferId": { "type": "string" },
                  "note": { "type": "string" },
                  "ttlMs": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Offer created" } }
      }
    },
    "/transport-requests/{requestId}/cancel": {
      "post": {
        "tags": ["negotiation"],
        "summary": "Cancel transport request during negotiation",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "requestId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Cancelled" }, "409": { "description": "Already converted" } }
      }
    },
    "/offers/{offerId}/counter": {
      "post": {
        "tags": ["negotiation"],
        "summary": "Counter an offer",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "offerId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "Counter created" } }
      }
    },
    "/offers/{offerId}/accept": {
      "post": {
        "tags": ["negotiation"],
        "summary": "Atomic accept → Booking (Idempotency-Key supported)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "offerId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "201": { "description": "Booking created" },
          "200": { "description": "Idempotent reuse" },
          "409": { "description": "Conflict / expired / double accept" }
        }
      }
    },
    "/offers/{offerId}/reject": {
      "post": {
        "tags": ["negotiation"],
        "summary": "Reject pending offer",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "offerId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Rejected" } }
      }
    },
    "/offers/{offerId}/withdraw": {
      "post": {
        "tags": ["negotiation"],
        "summary": "Provider withdraws pending offer",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "offerId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Withdrawn" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token from POST /auth/login"
      }
    }
  }
}